NativeJS Refactor
- moved error position locator to NativeJSEval
- extended NativeJSEvalResult type to include error
- updated documentation for runWrapper
