+++ src/features/nativeJS/NativeJSTypes.ts
@@ -1,3 +1,4 @@
+import { SourceError } from 'js-slang/dist/types';
 import { WorkspaceLocation } from 'src/commons/workspace/WorkspaceTypes';
 
 export const NATIVE_JS_RUN = 'NATIVE_JS_RUN';
@@ -7,6 +8,7 @@ export interface NativeJSEvalPayload {
 }
 
 export interface NativeJSEvalResult {
-  message: any;
   status: 'finished' | 'error';
+  value: any;
+  error?: SourceError;
 }
